requirements.txt
bs4
requests
pandas
numpy
scikit-learn
selenium
nltk
langdetect
transformers
datasets
flask
tensorflow==2.16.2
tf-keras==2.16.0
boto3
python-dotenv
sagemaker
docker